Qt: fixes for increments and evaluations on each keypress in textboxes
