+++ modules/qtwidgets/src/properties/doublevaluedragspinbox.cpp
@@ -50,7 +50,7 @@ namespace inviwo {
 DoubleValueDragSpinBox::DoubleValueDragSpinBox(QWidget* parent)
     : QWidget(parent)
     , spinBox_(new NumberLineEdit())
-    , valueDragger_(new ValueDragger<double>(spinBox_))
+    , valueDragger_(new ValueDragger(spinBox_))
     , invalid_(false) {
     setObjectName("valueDragSpinBox");
     spinBox_->setButtonSymbols(QAbstractSpinBox::NoButtons);
@@ -141,19 +141,6 @@ void DoubleValueDragSpinBox::selectAll() { spinBox_->selectAll(); }
 void DoubleValueDragSpinBox::stepDown() { spinBox_->stepDown(); }
 
 void DoubleValueDragSpinBox::stepUp() { spinBox_->stepUp(); }
-
-int DoubleValueDragSpinBox::spinnerDecimals(double value) const {
-    std::ostringstream buff;
-    utilqt::localizeStream(buff);
-    buff << value;
-    const std::string str(buff.str());
-    auto periodPosition = str.find(locale().decimalPoint().toLatin1());
-    if (periodPosition == std::string::npos) {
-        return 0;
-    } else {
-        return static_cast<int>(str.length() - periodPosition) - 1;
-    }
-}
 
 void DoubleValueDragSpinBox::setInvalid(bool invalid) {
     if (invalid_ == invalid) return;
